requirements.txt
numpy
pandas
pymupdf
pillow